Refactor landing page into seperate components
